third_party/blink/web_tests/platform/generic/fast/events/keydown-leftright-keys-expected.txt

This tests keyboard events with KeyLocationCode argument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 16
PASS lastKeyboardEvent.location is KeyboardEvent.DOM_KEY_LOCATION_LEFT
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 17
PASS lastKeyboardEvent.location is KeyboardEvent.DOM_KEY_LOCATION_LEFT
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 18
PASS lastKeyboardEvent.location is KeyboardEvent.DOM_KEY_LOCATION_LEFT
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 16
PASS lastKeyboardEvent.location is KeyboardEvent.DOM_KEY_LOCATION_RIGHT
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 17
PASS lastKeyboardEvent.location is KeyboardEvent.DOM_KEY_LOCATION_RIGHT
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 18
PASS lastKeyboardEvent.location is KeyboardEvent.DOM_KEY_LOCATION_RIGHT
PASS successfullyParsed is true

TEST COMPLETE
